+++ plugin/global/settings/settings.default.toml
@@ -25,17 +25,21 @@ CLICK_EDIT_BUTTON_TO_EXIT_INTERACTIVE_MODE = true
 # ctrl+click图表退出交互模式
 CTRL_CLICK_TO_EXIST_INTERACTIVE_MODE = false
 
-############### ArticleUploader ###############
-[ArticleUploader]
+############### article_uploader ###############
+[article_uploader]
 # 是否启用插件
 ENABLE = true
 # 插件名称
-NAME= "上传到所有平台"
+NAME = "文章上传"
 # 在右键菜单隐藏
-HIDE=true
+HIDE = true
 # 是否在右键菜单中可点击
-CLICKABLE = true  
-
+CLICKABLE = true
+# 快捷键配置
+UPLOAD_ALL_HOTKEY = "ctrl+shift+1"
+UPLOAD_WORDPRESS_HOTKEY = "ctrl+shift+2"
+UPLOAD_CNBLOG_HOTKEY = "ctrl+shift+3"
+UPLOAD_CSDN_HOTKEY = "ctrl+shift+4"
 
 ############### ArticleUploader ###############
 [ArticleUploader]
