+++ plugin/uploadArticle/ArticleUploader.js
@@ -1,10 +1,10 @@
 class ArticleUploader extends BasePlugin {
     init = () => {
         this.callArgs = [
-            {arg_name: "上传到csdn", arg_value: "upload_to_csdn"},
-            {arg_name: "上传到wordpress", arg_value: "upload_to_wordpress"},
-            {arg_name: "上传到博客园", arg_value: "upload_to_cn_blog"},
-            {arg_name: "全部上传", arg_value: "upload_to_all_site"},
+            {arg_name: "CSDN", arg_value: "upload_to_csdn"},
+            {arg_name: "WordPress", arg_value: "upload_to_wordpress"},
+            {arg_name: "博客园", arg_value: "upload_to_cn_blog"},
+            {arg_name: "上传到所有平台", arg_value: "upload_to_all_site"},
         ]
     }
 
@@ -22,7 +22,6 @@ class ArticleUploader extends BasePlugin {
             upload_to_cn_blog: "cnblog",
             upload_to_all_site: "all"
         }
-        console.log(type)
         const _type = map[type];
         if (!_type) return;
